backend/requirements.txt
django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.1.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
